requirements.txt
Flask
Werkzeug
marker-pdf
gunicorn
Pillow